requirements.txt
Flask==2.2.2
strawberry-graphql==0.144.1